kitchen/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kitchen"
version = "0.1.0"
description = "Reusable MLOps framework: data ingestion, experiment tracking, serving"
requires-python = ">=3.11"
dependencies = [
    # Data
    "dvc[s3]>=3.0",
    "pandas>=2.0",
    "pyarrow>=14.0",
    # ML
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    # Experiment tracking
    "mlflow>=2.10",
    "boto3>=1.34",
    # Serving
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "mangum>=0.17",
    # Monitoring
    "evidently>=0.4",
    # Orchestration
    "prefect>=2.19",
    # Data ingestion
    "kaggle>=1.6",
    # Params
    "pyyaml>=6.0",
]

[project.optional-dependencies]
serve = [
    "fastapi>=0.110",
    "mangum>=0.17",
]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "httpx",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["kitchen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=kitchen --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"
